address pr comments
remove extra line

address more comments
